Refactor #11: Made Maze an interface and RectangleMaze implements it
